(NOT WORKING) Adding edge detector class
Features :
- `class EdgeFinder` : implementation of the ACO interface to detect edge
- typo corrections

Missing :
- the class is not complete (at all)

+++ Sources/ACO_Interface.py
@@ -89,7 +89,7 @@ class ACO(object):
         """Returns the heuristic information for the edge between  the node `start` and `end`"""
         pass 
 
-    def _ApplyPolicy(self, current : int, adj : list) -> float:
+    def _ApplyPolicy(self, current : int, adj : list) -> int:
         """
         Determines the next node where to go for the calling ant. Parameters:
             - `current` : current node of the ant
